Remove what appears to be an unnecessary replace.
decode_rsa_key does a .replace('\\n', '\n') which replaces any literal
2 byte sequence '\n' with a newline character. The original commit that
added it is 38fd8514933bedaa5b23d7ba0b02c35353e3f05b but there's no
mention of why it's needed and there are no test cases that fail when
it's omitted. Any file that has a literal 2 byte sequence of '\\n' is
surely not a valid pkcs8 file so it seems perfectly valid to skip the
replace (and allocation) and return an error if one is encountered.

If it's determined that this check is needed for some reason please add
a unit test that explains why for future contributors.